+++ Metrics/App_Packages/HdrHistogram/AbstractHistogram.cs
@@ -1104,7 +1104,7 @@ namespace HdrHistogram
             while (recordedValuesIterator.hasNext())
             {
                 HistogramIterationValue iterationValue = recordedValuesIterator.next();
-                Double deviation = (medianEquivalentValue(iterationValue.getValueIteratedTo()) * 1.0) - mean;
+                double deviation = (medianEquivalentValue(iterationValue.getValueIteratedTo()) * 1.0) - mean;
                 geometric_deviation_total += (deviation * deviation) * iterationValue.getCountAddedInThisIterationStep();
             }
             double std_deviation = Math.Sqrt(geometric_deviation_total / getTotalCount());