+++ Metrics/Graphite/PickleGraphiteSender.cs
@@ -1,8 +1,7 @@
-﻿using System;
+using System;
 using System.Net;
 using System.Net.Sockets;
 
-using Metrics.Logging;
 using Metrics.Utils;
 
 namespace Metrics.Graphite
@@ -46,7 +45,7 @@ namespace Metrics.Graphite
                 {
                 }
                 client = null;
-                MetricsErrorHandler.Handle(x, "Error sending Pickled data to graphite endpoint " + host + ":" + port.ToString());
+                MetricsErrorHandler.Handle(x, $"Error sending Pickled data to graphite endpoint {host}:{port}");
             }
         }
 
@@ -67,7 +66,7 @@ namespace Metrics.Graphite
                 {
                 }
                 client = null;
-                MetricsErrorHandler.Handle(x, "Error sending Pickled data to graphite endpoint " + host + ":" + port.ToString());
+                MetricsErrorHandler.Handle(x, $"Error sending Pickled data to graphite endpoint {host}:{port}");
             }
         }
 
@@ -76,7 +75,6 @@ namespace Metrics.Graphite
             var endpoint = new IPEndPoint(HostResolver.Resolve(host), port);
             var client = new TcpClient();
             client.Connect(endpoint);
-            log.Debug(() => "Picked client for graphite initialized for " + host + ":" + port.ToString());
             return client;
         }
 
@@ -97,11 +95,8 @@ namespace Metrics.Graphite
             base.Dispose(disposing);
         }
 
-        private static readonly ILog log = LogProvider.GetCurrentClassLogger();
-
         private readonly string host;
         private readonly int port;
-
         private readonly int pickleJarSize;
 
         private TcpClient client;