+++ Metrics/Graphite/TcpGraphiteSender.cs
@@ -1,9 +1,8 @@
-﻿using System;
+using System;
 using System.Net;
 using System.Net.Sockets;
 using System.Text;
 
-using Metrics.Logging;
 using Metrics.Utils;
 
 namespace Metrics.Graphite
@@ -35,7 +34,7 @@ namespace Metrics.Graphite
                 {
                 }
                 client = null;
-                MetricsErrorHandler.Handle(x, "Error sending TCP data to graphite endpoint " + host + ":" + port.ToString());
+                MetricsErrorHandler.Handle(x, $"Error sending TCP data to graphite endpoint {host}:{port}");
             }
         }
 
@@ -51,7 +50,7 @@ namespace Metrics.Graphite
                 {
                 }
                 client = null;
-                MetricsErrorHandler.Handle(x, "Error sending TCP data to graphite endpoint " + host + ":" + port.ToString());
+                MetricsErrorHandler.Handle(x, $"Error sending TCP data to graphite endpoint {host}:{port}");
             }
         }
 
@@ -60,7 +59,6 @@ namespace Metrics.Graphite
             var endpoint = new IPEndPoint(HostResolver.Resolve(host), port);
             var client = new TcpClient();
             client.Connect(endpoint);
-            log.Debug(() => "TCP client for graphite initialized for " + host + ":" + port.ToString());
             return client;
         }
 
@@ -80,11 +78,8 @@ namespace Metrics.Graphite
             base.Dispose(disposing);
         }
 
-        private static readonly ILog log = LogProvider.GetCurrentClassLogger();
-
         private readonly string host;
         private readonly int port;
-
         private TcpClient client;
     }
 }