+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Remove ElasticSearch and InfluxDB reporters from the library.
 - Remove `MetricsConfig.WithHttpEndpoint()` feature from the library.
 - Remove [System.Configuration.ConfigurationManager](https://www.nuget.org/packages/System.Configuration.ConfigurationManager) dependency.
+- Remove LibLog. Use `Metric.Config.WithErrorHandler((e, msg) => {...})` directly if you need to log Metrics.NET errors.
 - Migrate tests from xUnit to NUnit.
 - Switch to SDK-style project format and dotnet core build tooling.
 - Use [Nerdbank.GitVersioning](https://github.com/AArnott/Nerdbank.GitVersioning) to automate generation of assembly